Make sure we honor externals, excluded and ignores in all pipelines
