+++ bin/index.js
@@ -6,12 +6,12 @@
 var browserify = require('browserify');
 var fs = require('fs');
 var mkdirp = require('mkdirp');
-var glob = require('glob');
 var xtend = require('xtend');
 var path = require('path');
 var async = require('async');
 
 var parseConfig = require('mendel-config');
+var applyExtraOptions = require('mendel-development/apply-extra-options');
 var mendelBrowserify = require('mendel-browserify');
 var mendelRequirify = require('mendel-requirify');
 
@@ -41,59 +41,7 @@ async.each(config.bundles, function(rawBundle, doneBundle) {
         });
     }
 
-    [].concat(bundle.ignore).filter(Boolean)
-        .forEach(function (i) {
-            b._pending ++;
-            glob(i, function (err, files) {
-                if (err) return b.emit('error', err);
-                if (files.length === 0) {
-                  b.ignore(i);
-                }
-                else {
-                  files.forEach(function (file) { b.ignore(file) });
-                }
-                if (--b._pending === 0) b.emit('_ready');
-            });
-        })
-    ;
-
-    [].concat(bundle.exclude).filter(Boolean)
-        .forEach(function (u) {
-            b.exclude(u);
-
-            b._pending ++;
-            glob(u, function (err, files) {
-                if (err) return b.emit('error', err);
-                files.forEach(function (file) { b.exclude(file) });
-                if (--b._pending === 0) b.emit('_ready');
-            });
-        })
-    ;
-
-    [].concat(bundle.external).filter(Boolean)
-        .forEach(function (x) {
-            var xs = splitOnColon(x);
-            if (xs.length === 2) {
-                add(xs[0], { expose: xs[1] });
-            }
-            else if (/\*/.test(x)) {
-                b.external(x);
-                b._pending ++;
-                glob(x, function (err, files) {
-                    files.forEach(function (file) {
-                        add(file, {});
-                    });
-                    if (--b._pending === 0) b.emit('_ready');
-                });
-            }
-            else add(x, {});
-
-            function add (x, opts) {
-                if (/^[\/.]/.test(x)) b.external(path.resolve(x), opts)
-                else b.external(x, opts)
-            }
-        })
-    ;
+    applyExtraOptions(b, bundle);
 
     if (entries) {
         // TODO: aync ../lib/resolve-dirs instead of hardcoded base
@@ -114,17 +62,4 @@ async.each(config.bundles, function(rawBundle, doneBundle) {
 function logObj(obj) {
   console.log(require('util').inspect(obj,false,null,true));
   return obj;
-}
-
-function splitOnColon (f) {
-    var pos = f.lastIndexOf(':');
-    if (pos == -1) {
-        return [f]; // No colon
-    } else {
-        if ((/[a-zA-Z]:[\\/]/.test(f)) && (pos == 1)){
-            return [f]; // Windows path and colon is part of drive name
-        } else {
-            return [f.substr(0, pos), f.substr(pos + 1)];
-        }
-    }
 }
